Setup minimal infrastructure to run scalafix.
I haven't actually runned the code, but it should look something like
that. The Scalafix API makes this way too hard, I really need to
refactor it into a simpler/smaller public API so that we don't have to
rely on scalafix.internal.

Currently only runs syntactic rules, to run semantic rules we should
setup a file watcher to listen to changes to .semanticdb files.