Index JDK sources.
The JDK needs to be special handled since it's included in the classpath
by default and does not appear in the `updateClassifiers` report from
sbt. The JDK on my machine contains ~2.5M loc but still indexes in ~8s
at a staggering rate of ~450k loc/s. The need to persist these indices
across editor sessions grows bigger.