Make loadFromTextDocument return schema.Database
This is just for consistency: the current implementation is "silly"
since it retrieves a semanticdb.Databse, converts it back to a
schema.Database and then the only thing we do is to convert it back
again to semanticdb.Database.
However this is readying the ground for future changes in
semanticdb-scalac, which may switch to returning schemas instead of the
high-level structures.