Run messages asynchronously instead of on the main thread.
Previously, we used Task.now to eagerly compute all values. This blocked
up the server immediately and made the server fail to handle requests
like shutdown.

This commit also reduces noise in the logs, which made it easier to
debug what was happening.  Previously we logged everything messages to
the client, including the fact that we were logging to the client. This
was a lot of noise and made it difficult to see what was happening.