Remove unnecessary -Dvscode.workspace argument.
The logs are emitted in `<working directory>/.metaserver/metaserver.log`
I didn't use the `directories` project to get the path because it
assumes you want a different location for config/cache and there is no
log-specific option. I personally want all metaserver files to be
bundled in the same directory so I can `rm -rf .metaserver`.