Run presentation compiler from single thread.
Previously, we ran the presentation compiler on the main scheduler which
defaults to a thread pool with 4 threads. While trying out the plugin
I noticed that the presentation compiler used to crash a lot with
cryptic error messages.
I suspect concurrent access to compiler data structures from semanticdb,
completions and hover caused these crashes. Let's see if this improves
the situation. I tested this change manually by running the plugin
for a small session and didn't see any errors in the logs.