Integrate with sbt language server.
This commit adds an integration to connect to sbt 1.1 language server
to compile via sbt on file save and publish error messages as
diagnostics. We reuse the LanguageServer and LanguageClient
utilities but we are now a client instead of server, compared to when
connecting to editors.

This commit adds a new SbtConnect command to re-estabilish a connection
with sbt in case of a disconnection, for example if you exit sbt and
open again. Ideally, we can figure this out automatically in the future,
but a manual approach is a good start.

There are two settings to control this new feature

- sbt.enabled, must be true to use this feature, false by default.
- sbt.command, which command to execute on file save. This makes it
  possible to for example run or test instead of compile or select
  a subproject in a multi module build.

I did not upgrade the build to sbt 1.1 because intellij import doesn't
work with sbt 1.1 yet for some reason.