Create mtags module with a complete solution for navigation.
This commit introduces a new module mtags that implements goto
definition with a new design that achieves the following goals:

- source indexing is fast, Scala sources are indexed at ~800k loc/s
  and Java sources don't require any indexing altogether.
- the in-memory index uses little ram, an index for ~1.5 million lines
  of Scala takes ~15MB.

The new symbol indexer is able to handle different kinds of sources:
source files, sources.jar and SemanticDB files.  Source and  SemanticDB
can complement each other, the indexer falls back source indexing when
the SemanticDB file is either missing or is incomplete due to for
example macro annotations.

An important difference from the old indexer is that the new indexer is
not responsible for computing the range position inside the source file.
The final range position is left for a separate module that handles
other concerns such as stale SemanticDBs using edit distance.