Refine sbt build detection
The previous logic assumed all sbt builds have `build.sbt` but that is
not correct. This commit changes the logic to guard against
`project/build.properties` with a `sbt.version` key. This key is
used by the sbt launcher and the sbt launcher generates this file
when `project/build.properties` is missing.