Improve sbt launcher.
This commit improves how the `sbt bloopInstall` step reproduces
the users `sbt` command.

- Respect .sbtopts and .jvmopts
- Use fully-qualified `java` binary
- Allow users to pass in custom sbt JVM options via
  `workspace/didChangeConfiguration` or system properties.

We borrow some code from IntelliJ Scala to ensure our handling of
`.sbtopts` is not a blind guess of how it's interpreted normally.

We don't system process to an `sbt` command because that failed CI on
Windows. We might end up having to go that route if this doesn't work
but I want to first try and use the same approach as IntelliJ here.