Document installation steps for Emacs.
I'm not too happy with the Emacs lsp-mode client. Most critically,
it doesn't seem to handle diagnostics for multiple files making
it difficult to fix all the compile errors in larger project.

Either way, this is a good start having Emacs documented on the website
will make it easier to upgrade everyone to new releases of Metals and
lsp-mode.