Move critical file watching in-house.
Previously, we registered on file watching pattern `**/*.{scala,java}`
to the editor if the editor supported file watching.  For editors that
didn't support file wathching we used our own custom file watching
solution which targeted precise directories in the build.

This commit removes the registration of the `**/*.scala` pattern so that
we use the in-house file watching solution for all editors.  The benefit
of this change is that we have a single implemementation that we know
works the same for all editors.

However, we still use the editor file watching support (if available)
for nice-to-have but non-critical patterns like
`$workspace/project/*.sbt`. Our in-house implementation only supports
recursive file watching so it's not easy to watch those patterns
ourselves.

Sidenote: the LSP spec does not specify the syntax for targeting precise
directories for dynamic file watching registration. My experiments show
that the syntax supported by VSCode requires is weird on Windows, it
uses backward slashes until the workspace directory, and forward slashes
after that:
```
C:\Users\Bar\workspace/project/*.scala
```