Implement fast and low-overhead find references.
To respond to a "find references" request, we iterate through all
semanticdb files in the workspace and query an in-memory bloom filter
for that semanticdb files whether it references a given symbol. If yes,
we read the semanticdb file from disk.

This approach is memory efficient, for a project like Prisma the index
uses ~1.5Mb to index ~60k symbol references at around 22 bytes per
symbol reference. Most reference requests respond on within 10ms, even
when there are several thousand references like for `String`.