Respect workspace folders for Scalafmt formatting, fixes #509.
Previously, we only looked for `.scalafmt.conf` in the `rootUri`
directory. After this fix, we fallback to `workspaceFolders` if we don't
find a `.scalafmt.conf` file. In the case of formatting, it's not big
trouble to support multiple workspace roots. In the general case it may
be tricker, for example I have no idea how we would handle two separate
sbt builds. We can take it case-by-case.