Implement completions that turn string literals into interpolators.
This feature stresses several features of the LSP completion API and it
seems that different editors interpret them differntly. This commit
works great in VS Code but it produces broke code in Atom. We can
disable these custom completions for editors that don't implement the
full completion API or interpret the TextMate snippet syntax
incorrectly.