Insert local import when completing workspace symbol.
Previously, Metals inserted a fully qualified name at the completion
position. With this commit, we now insert a local import in the nearest
enclosing block. Down the road, we can add a code action to "Move import
to global scope" to clean up local imports inserted by Metals.

Other improvements in this commit:

- we render the package name in the completion item label so that it's
  more visible what you are importing.
- we no longer use insert text, which is a deprected LSP feature and
  has unclear semantics between editors. Instead, we only use `textEdit`.