Add option to render `override def` as unicode 🔼
The "override def" string takes up a lot of horizontal space in the
completion label. It looks pretty good with unicode 🔼 instead but
since I'm not sure that renders nicely in all editors we keep ascii
"override def" as the default with a server property to opt-into
rendering it as unicode, which is enabled by default only for vscode.