Sort symbols defined in the current file to the top, fixes #618.
Previously, we sorted global symbols defined in the current file in the
same group as global symbols from the root scope. Now, we priotize
symbols in the current file.

This change has the nice side-effect that `::` no longer appears at the
top of all completions in an empty scope without a prefix.