Always widen type for hover.
Previously, hover frequently showed unwanted singleton types.  Now, we
always widen the type so that it's more readable. This change may cause
unwanted widening in some cases, but we can deal with that if some
reports a reproduction. I think it's better to guard against cases we
don't want widening instead of guarding against cases we want widening.