Don't escape dollars with trailing letter character.
Previously, we escaped all dollars in string literals into `$$` to
preserve the meaning of the string after selecting the completion.
Although this is a technically correct approach, it can be an annoying
behavior in practice when you type out a string like
```s
"Hello $name, this is $mess@@"
```
and you intended to `$name` to be spliced but it was escaped instead
into `$$name`.

Now we don't scape the dollar if it's followed by a letter character
under the assumption that's supposed to be a splice.  Presumably, the
most common usage of dollars in string interpolators is to splice
variables so behavior will more often align with the user's intention.