Generate exhaustive pattern match for sealed types.
Previously, you had to write on case at a time when pattern matching on
a sealed type. Now, you can complete on the `match` keyword and select
the "match (exhaustive)" completion item to generate all cases.  This
commit also introduces a new "match" completion (non-exhaustive) that
generates a single case and triggers a new completion at the case word,
for editors that support `CompletionItem.command`.