Make tests pass with the new goto definition fallback.
Our test suite stressed several invariants for goto definition that
broke once we started using the presentation compiler as a fallback.
This commit improves the goto definition fallback to make most of
the invariant tests pass, and updates several expected outputs in the
tests where the invariants no longer long (for example, definition and
references are no longer a bijection with dirty buffers.