Set timeout for shutdown procedure, fixes #715.
Previously, it was possible for the Metals process to never exit if the
shutdown future never returns.  Now, we always call `System.exit` even
when the shutdown runs forever.  Without a reliable reproduction, it's
difficult to confirm 100% that this change fixes issue #715. However,
this change should close one potential reason for why the process doesn't exit.