replace java streams with generators
Previously, we created a wrapper for file walking and children listing
to automatically close the underlying java stream, but it lacked the
flexibility and ease of use of standard collections.
Now, we use the geny library to close the stream automatically. It also provides a
subset of scala collection methods (forall, map, etc.)