Don't file watch inside jar files.
The Scala compiler supports a feature where it compiles the output
classfiles directly to `*.jar` files. This feature is used in build
tools like Pants and Bazel.

Previously, Metals would crash on build targets that used this feature
of the compiler. Now, we don't do anything even if it means missing file
watching notifications that are needed for the "find symbol references"
index.