Avoid race conditions during slow/test command.
See https://github.com/sbt/sbt-buildinfo/issues/152 for a discussion on
how sbt-buildinfo uses the same target directory for compiling a project
across different Scala versions. Sharing the target directory is safe
under normal usage of the sbt task engine but it's risky when using
`State`. An example of a failed CI job caused by this bug is
https://travis-ci.org/scalameta/metals/jobs/590512863

This commit reduces the risk of race conditions during sbt-buildinfo
code generation by forcing the `publishLocal` steps for Scala 2.11 and
Scala 2.13 to run synchronously, instead of in parallel.