Fix Runtime Error In Metals Bench
A path to file needed to run the `CompletionBench` benchmark was hard coded as scala `2.12.8`. Since this is no longer the version of Scala used to compile `metals` this fails at runtime.

This was changed to use sbt-buildinfo to populate the correct version of scala for the given build from the build definition directly.