Disable EOL Conversion On /.jvmopts
Add a `.gitattributes` file to prevent end of line (EOL) conversion on the `/.jvmopts` file. This was the causing windows based CI failures as the JVM options were being parsed with a carriage return at the end of them.

This commit also re-enables use of the `.jvmopts` file in the Windows CI environment.