Respect library excludes in Pants BUILD files
Previously, we included all transitive library dependencies on the
classpath of a Bloop target. This behavior caused runtime
`NoSuchMethodException` from binary conflicting jars that were supposed
to be excluded from the classpath. Now, we don't add jars to the
classpath if they are excluded in the Pants BUILD file fixing those
runtime exceptions.