Don't generate synthetic project root project if it conflicts.
Previously, if the root directory contained a BUILD file then the export
step would generate both targets for the BUILD file as well as a
synthetic "*-project-root" target. This caused problems in IntelliJ
where it would remove one of the targets because they have a conflicting
source directory. Now, we don't generate synthetic projects if there is
already a project which has that directory covered.