Make use of latest Pants improvements
Previously, the `fastpass` command-line was compatible with most recent
versions of Pants. Now, it requires Pants v1.24 or newer. The motivation
for this change was to take advantage of the new
`./pants export-dep-as-jar` goal that was added specifically to support
an exporter tool like fastpass.

The benefits of this changes include:

* faster export times
* sources.jar for non-exported targets fixing navigation bugs
* simpler logic overall

Additionally, this commit upgrades to the latest Metaconfig in order to
support tab completions.