Add support to query remote language server as fallback.
Previously, Metals only supported responding to definition/requests from
a local index that was built by Metals. For some large codebases it's
not practical to index all the code locally making it difficult to use
Metals.

Now, Metals supports querying a remote language server as a fallback
when the local index is not able to respond to definition/references
requests. This feature is especially helpful in a large monorepo where
there exists a lot of code but the user is only actively working on a
small part of the codebase. With a remote language server, users can
query the remote language server for source files that the user is not
editing without building a local index.