Support Pants' `strict_deps`
Previously, `fastpass` was not calling Pants with
`--respect-strict-deps`, which means that Pants will not respect
`strict_deps` when doing the export.

The output of Pants' `export-dep-as-jar` also changed to include a clear
separation between compile-time and runtime classpaths.

This commit updates fastpass so that `--respect-strict-deps` is passed
to pants, and so that fastpass can understand the new output of
`export-dep-as-jar`. The resulting Bloop configuration files will
include the correct compile-time and runtime classpaths.