Remove the declaration of the experimental features in initialize.
Previously these use to be set so that during the intiliaze phase a client could
check to see if a server supported something. Since we have now moved away from
using the clientExerimentalFeatures it makes sense to no longer declare these. This
doesn't mean that aren't able to be set, just that we aren't explicit to the clients
that we support them anymore. Plus these are out of date anyways.