Add in spacing in vscode icons only.
Since VS Code is the only one using an actual icon for spacing it's
better to put the spacing there since we are doing a replace on the
icon and times producing labels that have one empty space and other
times two empty spaces causing odd spacing in other clients. This
change will allow for more consistent spacing in those clients while
keeping the same spacing in VS Code.