Add in explicit `TextDocumentSyncOptions`.
Up to this point we would just return a 1 for this. This sort of leaves
it up to the client how they should handle this then. According to the
spec, it actually seems that if there isn't a specific key in the
`TextDocumentSyncOptions` sent in, then it should be set to false. One
of those options is `save`, which is in control of sending
`textDocument/didSave` events. Since Metals requires this for things
like worksheets, we should make sure we are explicit about us requiring
it to ensure clients set this correctly.