Set BuildTool version via BuildInfo.
Previously we set the recommended version for the various BuildTools
by having the version hardcoded in the BuildTool file. This moves
the version for Maven, Mill, and Gradle into the build.sbt file and
uses BuildInfo to access those versions like we currently do with
the SBT build tool. This will also allow us to get some help from our
friend ScalaSteward.