Remove gson runtime reflection with manual reads for configuration.
While this is a bit more verbose than what we had, it allows this to be
much more explicit and represent what we actually want to do. When this
was first done (https://github.com/scalameta/metals/pull/1674) I had a
bunch of issues with `null` being set for values that didn't exist, not
easily being able to use `Option` in the structure, and also needing
default constructors in order for it to work correctly. I didn't love
it, and have wanted to revisit this for a while. I also figured this was
good to tackle before diving into
https://github.com/scalameta/metals/issues/1895.

By doing it the way we previously did it caused us to have to check
configuration like this:

 1. Is `InitializationOptions` set to a non-default value?
 2. Is `ClientExperimentalCapabilities` set to a non-default value?
 3. If so, take their value or use the default set in the initial
    config.

Instead, all we really wanted to know was really:

  1. Did `InitializationOptions` set anything?
  2. Did `ClientExperimentalCapabilities` set anything?
  3. If neither, _then_ take the default

Again, this is a bit more verbose but I think ultimately more clearer
and closer to what we actually want.