Add option to not install and apply bloop in gradle builds
Users might want to enable bloop manually in their builds or might use a
Gradle distribution that uses and configured Bloop by default.

The following commit adds a way to tell Metals that some builds are
already configured and don't require the init script to install and
configure bloop in every project of the build. As a result, Metals only
runs `bloopInstall`.