Add in , as a trigger for SignatureHelp.
This will allow users to enter a parameter or two, step out, and then back
in and still get signature help. Currently, if you leave the parameters
and step back in, you will no longer get signature help. Also some clients
won't update and send another request to the server while you are in the
parameter group and hit ,. Now with this being explicit they will.