Create ExcludedPackageHandler to handle exclusions.
The idea is to have all the logic included in one place to handle
exclusions both from the base exclusions and also from the user input.
This also adds in a way to remove an item from the default list by prepending
it with `--`. Then the class just exposes a function to check pkgs in various
places when indexing is happening.