Fix missing classes from _empty_ package in treeView.
Previously if we were looking for memebers underneath _empty_ we ended up
looking for them in <uri>/_empty_ rather than just looking for them in the
root where they are stored. This adds in another check where when we look to
see if a symbol is in the root package, we also check to see if it's an empty
package, and if so, then just return the root value to look for it.