Don't search for runTest code lenses in Ammonite scripts and worksheets.
Up to this point we search for all types of code lenses even in Ammonite
scripts and in worksheets where there won't be any run or test code
lenses that appear. This will save us a bit as we won't look through all
the symbols in the textDocument for the main method.