Change new-scala-file to enable quick creation of a file.
This changed the current functionality of the new-scala-file command to allow
for another argument in the third position, which signifies the file type. Previously,
we allowed for the name of the file to be in the args in the second position, but now
we also allow the file type to allow for a client to create a file with a single command.
This mainly will allow for the quick-creation of worksheets which can be bound to an
editor action or command.

This also changes up the NewFilesLspSuite to cover the other arguments. Previously the
functionality of the name being passed in wasn't covered. Now, that functionality and
also providing the fileType is also covered in the tests.

Ideally the args for this command would have been an JSON object allowing for the necessary
keys to be set or not, but since it wasn't that way, I hesitate to change it now since
some clients may be using the array with 2 items (directory and name). Implimenting it
this will will allow for backwards compatability with any clients that may already be
utilizing the command with both directory and name being passed in.