Add java sources to standalone worksheets and compile before evaluation if needed.
This adds in java sources to the StandaloneSymbolSearch in order to
be able to correctly goto definition in Java sources when you have a
standalone worksheet.

This also adds in a check to see if your worksheet belongs to a build
target before evaluating. If it does belong, then it compiles the target
first.