Improve the process where a user has already chosen sbt server.
This commit takes a closer look at the second situation where a user has already
started sbt in the past, which ensures that a `.bsp/sbt.json` already exists and
that't they've triggerd the `build-server-start` with sbt, saving the choice in
the db. Upon starting Metals again, this will automatically be detected, and sbt
will be started and a bsp connection automatically made.