Capture BuildServerCapabilities and refactor `workspace/reload`.
Previously we weren't really capturing the capabilities. Since now we need to check
to see if the server supports reload requests, we carry it forward and check against
it later.

The `workspace/reload` refactor basically just allows now for all connections to be
reload, not just the main one. Plus it captures the error that the CompletableFuture
may have to display to the user that they tried to reload a workspace that isn't working.