Refactor out SbtServer and create BuildServerProvider.
Before this we have a SbtServer class to handle the initial and running and connecting
of the sbt BSP process. Instead now, we have a BuildServerProvider that is meant to be
mixed in to a BuildToold and provide the necessary functionality to generate a BSP
config file, and then connect to it.