Add type decorations for defintions
Previously, we added inferred type for synthetic types, however it's also possible to use semanticdb to get type of definitions.

To that end in this PR showing type was implemented in the following steps:
1. Trees for the file are taken from the cache.
2. Tree is searched for any definitions that do not have a declared type and ranges for those are returned.
2a. If the definition is a method we additionally calculate the position for the type to be shown.
3. We go through occurences and if a symbol is a definition and range was found previously to not contain declared type, the new decoration is added.
3a. For methods we use the position previously calculated to show the type.

A better type printing for tuples and functions was also added to show:
- `(Int, String)` instead of `Tuple2[Int, String]`
- `(Int) => String` instead of `Function1[Int, String]`

Additionally, the new feature will work with any changes to the file including ones that break the parsing of the file.