Add in Scala Sources to standalone worksheets.
Like we did to fix #2133, if the user is in a situation where they are
given a standalone compiler for worksheets, we then ensure that they not
only have Java sources available to them, but also Scala sources. We
default to the version of Scala in Metals itself.

I do have a check in there to make sure there is no "scala-library" in
sources, but that _probably_ isn't necessary as (in this scenario) the
only time that would happen would be if they have included it as an
import in their worksheet. However, just to be safe, I added it in.

I also moved the logic for this from the compilers to the
StandaloneSymbolSearch apply methods to ensure that we don't get hit on
this in the future if we use these elsewhere.

Closes #2266