Fix auto-connect BSP server without explicit choice.
Before this if a `.bsp/sbt.json` file existed and there was no `.bloop/`
yet then it would attempt to auto-connect to sbt. This behavior changes
this slightly now so that the behavior I just explains _only_ happens
if an explicit choice was made before, or if there is a `.bsp/<build-server>.json`
entry that is a Build Tool that Metals doesn't have Bloop support for.