Bump version of organizeImportRule to latest
Without this being the latest users may be using certain new
settings that have been added which will cause a warning that
the rule can't be found, but it's just because the setting
doesn't yet exist with this version.

I found this when I found #2306, but this doesn't fix the the
underlying issue.