Add support for Scala 3.0.0-M3
In order not to deprecate 0.27.0-RC1 and 3.0.0-M1 we needed to use an older munit version for those, as we needed to drop support for them in munit. There was also small needed fix, since typeRef was no longer available, but paramRef does invoke it underneath. Another issues that we faced was the fact the mtags no longer compiled on the 0.4.5 dotty plugin version and we needed to update both it and sbt, which will cause some tests to fail, which we ignored.