Refactor doctor and make sure issues with sbt are properly reported
The main goal of this refactor is to make sure we have just one category of problems that will be reported both in the message request and the doctor. Previously, both would need to be updated independently and testing for those was scarce. Thanks to having a single category of problems we can now properly unit test them to make sure every message is sensible. It's also easier now to see what each problem is caused by and what message will be displayed to the user.