Calculate semanticdb for standalone files
This solves all issues connected to semanticdb in standalone files. It's now possible to find references, rename etc in standalone files. This was also helpful for testing out and fixing an issue with URIs being wongly produced for semanticdb files produced in the compiler.

In the course of implementing this it also turned out that dependency sources would not be compiled properly if they came from the compiler or from a worksheet.