Remove Atom from the docs.
While it's sort of sad to do this, it's probably time to remove
Atom from the supported editors list in the docs. Since the plugin
is no longer maintained, Atom usage is lowering every year, and
the most recent survey had the lowest support shown for Atom, it's
probably as good as time as any.

Good times Atom.