Grab the newest semanticdbVersion.
Since we want this value to be able to be changed from the user,
but we also want to ensure it's the newest, we still allow for them
to set this, and we will use it if it's the newest, but if not, we
default to the semanticdb version that we have set in Metals. Normally
this value will never have to be touched, but it covers the case where
we haven't released support for a specific version yet and someone needs
to bump the semanticdb version locally.