Search also transitive sources when falling back from symbol search
It seems in cases where we have `package object myObject {}` inside `myObject.scala` file the path given to us will be `myObject/myObject`, but the actual symbol is `myObject/package.`. This normally works with a standard  naming where package objects are put into `package.scala` file, so the path would be `myObject/package`.

As a fallback for symbol search we also check workspace sources for the file, but we haven't done it transitively, so files in dependency project would not be found. This fixes it so that the fallback works in those cases too.