Ensure Bloop isn't offered as a server in an unsupported workspace.
Previously if a user would issue a `bsp-switch` command and there was an
entry in `.bsp/` for a server, then no matter what, Bloop was offered as
a choice. However, there may be a situation where a workspace is a
workspace that does have a `.bsp/` entry, but the build tools is one
that Bloop doesn't work with. In this scenario, Bloop shouldn't be
offered since it doesn't know how to import from this build tool.

An example. Let's say we have build tool `xyz` and server `xyz` with a
corresponding `.bsp/xyz.json` file. If a user open Metals in this
workspace, it will auto-connect to `xyz`. If the user issues a
`bsp-switch`, they will just see a warning now that says:

> Unable to switch build server since there is only one supported build
> server 'xyz' detected for this workspace.

Fixes #2424