Remove dependency on io.methvin:directory-watcher.
As of https://github.com/scalameta/metals/pull/2014 we replaced the
directory watching functionality with that of Swoval. However, we do
re-use some of the functionality of directory-watcher, mainly the
DirectoryChangeEvent and some hashing utils that it utilizes. This
change removes the full directory-watcher dependency but copies over the
few necessary classes from that library into a
scala.meta.internal.watcher package.

Closes #2365