Pass URI to `semanticdbTextDocument` method
Previously we would use string for that, which is not neccessary as we need to turn it back into URI anyway. This seems to also cause possibly cause the issues with:
```
Feb 01, 2021 11:32:01 AM scala.meta.internal.pc.CompilerAccess handleError
SEVERE: null
java.nio.file.ProviderMismatchException
        at sun.nio.fs.UnixPath.toUnixPath(UnixPath.java:200)
        at sun.nio.fs.UnixPath.relativize(UnixPath.java:410)
        at sun.nio.fs.UnixPath.relativize(UnixPath.java:43)
        at scala.meta.internal.pc.SemanticdbTextDocumentProvider.$anonfun$textDocument$2(SemanticdbTextDocumentProvider.scala:38)
        at scala.Option.map(Option.scala:230)
        at scala.meta.internal.pc.SemanticdbTextDocumentProvider.textDocument(SemanticdbTextDocumentProvider.scala:37)
```
I also added a Try there just to be on the safe side and additional check when calculating semanticdb in InteractiveSemanticdb.