Introduce a `metals/buildTargets` LSP extension.
You can find more of my mindset behind this in this discussion:
https://github.com/scalameta/metals/discussions/2479, but I'd like to
propose that we expose this info. One concrete thing that I'm planning
on using this for is having `nvim-metals` be able to quickly get the
build targets for a workspace, which then the user can just quickly
choose the one they want, and that is how the debug process or run
process starts without the need to manually add the build target or rely
on code lenses.

After playing around with this a bit, I'm also quite excited about other
possibilities that this opens up. There is _so much_ information in here
that client extensions can use. Everything from simple things like
`javaHome` or Scala versions to more complex things like the full
classpath for a build target can be accessed by extensions to do all
sorts of things. I'm excited to play around with this more.