Don't populate completion detail if it's an empty string.
This probably isn't that important, but I noticed this recently and had
to fix it in a completion extension
https://github.com/hrsh7th/nvim-compe/pull/177.

The detail field is optional in the `CompletionItem` so it might be best
to just not set it if the detail in `""`. If not, it may cause weird
things like showing an empty resolve float if the client doesn't have
the logic to filter that out.