Fallback to Scala3 dialects when parsing library sources
To enable finding symbols in library sources we use the parser to find the exact symbol that we need. Previously it would not work for Scala 3 due to using the Scala213 dialect. Now, we fallback to Scala3 dialect in case of issues, which is a temporary workaround until https://github.com/scalameta/metals/issues/2493 is fixed.

I tried to do it here, but it turns out it's a much more compelx issue and would not want to merge it before the release.