Fix standalone Scala 3 worksheets
In one of the previous PRs we changed the parser to take into account the real dialect of the code to parse based on the build target the file is located in. Recently, we also added the possibility to change the default Scala version, which allowed users to use standalone Scala 3 worksheets. We haven't however changed the parser dialect that is used for those worksheets, which results in Scala 3 specific code to be marked as an error. In this PR I am making the parser use the correct parser dialect.