Fix rename/references for worksheets
Thanks to the recent work on generating semanticdb for sbt, it's also possible to generate semanticdb for worksheets and in this way allow:
- references
- rename
- document highlight

The only thing missing is `Go to implementations` which doesn't work for anything that is not indexed globally. We don't want to index worksheets globally as their symbols are local to the file they exist in.