Try to capture a better message on BSP exceptions.
So in some cases if it was the reload that would fail, `T` here would
just be `Object`, which isn't very useful to show that it failed in
attempting to get `T`. Also, we are sort of swallowing the actual
message, so this just adds in a check to see if it's `Object`, then just
skip it and return the `Throwable` like normal, and if `T` is something
useful like `CompileResult` then we capture that _and_ also the original
`Throwable` message to pass on and later log.