Fix renaming named parameters in constructors
Fixes https://github.com/scalameta/metals/issues/1086

This however will not work the other way. Main parameter will not be renamed when rename is started from the named parameter. That needs to be fixed in https://github.com/scalameta/metals/issues/1553

At first I tried to change the occurrence of the named parameter inside scemanticdb, that however proved to be more difficult and I was not 100% sure if the current status is wrong. Since it involves quite a bit of changes for the other tools I decided against going that way.