Add init option to strip color from DAP communication.
This is meant to address the situations where a DAP client may not be
able to handle color codes in the DAP output. Instead of trying to build
this into both sbt and Bloop I think it's just easier to strip it here
and put it behind a flag. The normal behavior is unchanged, but now a
client can set `stripDAPOutputColor = true` and get the color codes
stripped.

Closes #2061