SymbolIndex: multiply symbol locations
This change fixes a set of problems(go to definition / references) related to the existence of the same symbol in different locations.
In particular: fixes #2265

Also, the way how dependency sources are extracted into `.readonly` was changed.
Now, if the source comes from source-jar it will be saved in: `.metals/readonly/dependecies/${source-jar-name}`.
This fixes navigation when there are several versions of the same library across build targets.

Co-authored-by: Chris Kipp <ckipp@pm.me>