Fix issues with old typesafe config in the worksheets
It seems that somehow when using akka in worksheets the actual typesafe config lib that is being loaded is coming from mdoc, not from akka. Which means that akka will not work and throw ClassNotFound exceptions since mdoc has 1.2.1 version on the classpath. We should also fix it in mdoc, but I have currently no idea how to fix it (we should be using the correct classpath currently), so I decided to make the classpath a bit thiner as worksheets do not need everything that mdoc brings. Especially, as there are no config options, so no need for the typesafe config.

Fixes https://github.com/scalameta/metals/issues/2525 hopefully