Surface error to user when they pass in an invalid envFile.
While probably not that common, this change simply surfaces an error to
the user if they happen to pass in an invalid `envFile` while trying to
test. In my case I hit on this because I removed my `envFile` entry
but left the key sending in a `""` as the path. It blew up in the logs,
but didn't actual show me anything on the client side. This will help in
that situation.