Add support for Scala 2.13.6
Scala 2.13.6 also includes the support for using some of the Scala 3 syntax. When the proper options is detected "-Xsource:3", we now change the dialect to be able to parse the enriched syntax. Additionally, we also now make it the default in case we don't know anything about the scalac options.