Print better types on hover
This improves a couple of things:
- show shorter type if it's importable (also needed further on)
- when hovering on package it's full path is shown (the change was needed due to the previous one)
- show final modifier
- show enum and case instead of final object and final val

The first change was needed also for properly writing types when doing inferred type action or overrides. The other changes stemmed from that. I think we can still further improve the hover information later on, but I don't want to do everything in one PR