Add support for textDocument/selectionRange.
Note that the way this is implemented and test takes a bit of liberty
mainly because I'm hitting on (what I would call) unclarity in the spec
about how this is exactly supposed to work. The interesting part is that
according to the
[spec](https://microsoft.github.io/language-server-protocol/specifications/specification-current/#textDocument_selectionRange),
the client can send in multiple positions in the `selectionRangeParams`.
From this point, things get interesting. I'll outline the two editors
that I was testing in because they behave pretty differently even though
the result is pretty much the same.

The steps it follows:
- Grabs the position your cursor is in, and sends that in as a single
  position
- The selection range is returned and each range has the parent filled
  in. So you get the entire tree
- When you continue the selection range, no further requests are made to
  the server. This is actually all done with one request, and it seems
  that if the full tree is there, then it just skips making any
  additional requests

This works pretty well and is actually pretty great that no other
requests are necessary. However, it then begs the question of when
should VS Code ever send in more than a single position for this
request? I actually have no clue. Even selecting a range and triggering
it from the range has VS Code sending in the start of the range, not the
entire range or multiple positions.

So contrast this with coc-metals utilizing coc.nvim

So while this looks very similar, it's radically different in the LSP
communication.

The steps it follows:
- Grabs the position your cursor is in, and sends that in as a single
  position
- The selection range is returned and each range has the parent filled
  in. So you get the entire tree
- When you continue the selection range an entire new request is sent,
  this time with first returned range split. Meaning the first position
  in the array was the start position of the returned range, and the
  second positions is the end position in the returned range
- Now there are two different selection ranges returned. In some
  scenarios (most scenarios from what I can tell) the actually return
  the same tree since the closes enclosing tree is the same whether you
  approach it from the start or the end. - It somehow intelligently
  skips the first enclosing tree if it's the same that was highlighted
  before and then continues on.
- Again, the further you expand, new requests are sent every time, but
  never containing more than 2 positions.

If you watch closely, you'll notice a very slight difference in the VS
Code ranges where it has one more selection range that goes to the
beginning of the line. That's actually not part of the selection range
returned, so I have no idea why it does that. It's almost seems like
they are doing a bit more behind the scenes than the spec outlines, but
still mainly following the full tree that is returned.

It doesn't bother me that much since this works pretty well in both
editors, but it matters in how we test this. I'm a bit stuck at the
moment how to test the step after the initial expansion since there
never is a next step in VS Code, and in coc.nvim it also seems to be
doing some stuff behind the scenes to choose the next range correctly.
I've a snippet of trace code in the test file if you're curious what it
looks like.

Therefore I took the liberty of testing this similarly to how jdtls
does.
[eclipse.jdt.ls](https://github.com/eclipse/eclipse.jdt.ls/blob/64b15c5a9e5b11f62ceb5163ceb6930d5dea7129/org.eclipse.jdt.ls.tests/src/org/eclipse/jdt/ls/core/internal/handlers/SelectionRangeHandlerTest.java)
it looks like they just pass in a single position and then ensures all
the positions in the selection range are as expected. The tests here do
the same.

Closes #2861